+++ qgis/providers/ogr/qgsogrprovider.h
@@ -59,19 +59,19 @@ class QgsOgrProvider:public QgsVectorDataProvider
      * Get the first feature resulting from a select operation
      * @return QgsFeature
      */
-    QgsFeature *getFirstFeature(bool fetchAttributes = false, int dataSourceLayerNum = 0);
+    QgsFeature *getFirstFeature(bool fetchAttributes = false, size_t dataSourceLayerNum = 0);
 
     /** 
      * Get the next feature resutling from a select operation
      * @return QgsFeature
      */
-    QgsFeature *getNextFeature(bool fetchAttributes = false, int dataSourceLayerNum = 0);
+    QgsFeature *getNextFeature(bool fetchAttributes = false, size_t dataSourceLayerNum = 0);
 
     /**Get the next feature resulting from a select operation.
     *@param attlist a list containing the indexes of the attribute fields to copy
     *@param getnotcommited flag indicating if not commited features should be returned
     */
-    QgsFeature *getNextFeature(std::list<int> const& attlist, int featureQueueSize = 1, int dataSourceLayerNum = 0);
+    QgsFeature *getNextFeature(std::list<int> const& attlist, int featureQueueSize = 1, size_t dataSourceLayerNum = 0);
 
     /** 
      * Get the next feature resutling from a select operation
@@ -79,7 +79,7 @@ class QgsOgrProvider:public QgsVectorDataProvider
      * that the feature is valid. Use QgsFeature::isValid() to check
      * the validity of a feature before using it.
      */
-    bool getNextFeature(QgsFeature &feature, bool fetchAttributes = false, int dataSourceLayerNum = 0);
+    bool getNextFeature(QgsFeature &feature, bool fetchAttributes = false, size_t dataSourceLayerNum = 0);
 
     /** Get the feature type. This corresponds to 
       WKBPoint,
@@ -90,7 +90,7 @@ class QgsOgrProvider:public QgsVectorDataProvider
       WKBMultiPolygon
      * as defined in qgis.h
      */
-    int geometryType(int dataSourceLayerNum = 0) const;
+    int geometryType(size_t dataSourceLayerNum = 0) const;
 
     /** return the number of layers for the current data source
 
@@ -103,18 +103,18 @@ class QgsOgrProvider:public QgsVectorDataProvider
     /** 
      * Get the number of features in the layer
      */
-    long featureCount(int dataSourceLayerNum = 0) const;
+    long featureCount(size_t dataSourceLayerNum = 0) const;
     /** 
      * Get the number of fields in the layer
      */
-    int fieldCount(int dataSourceLayerNum = 0) const;
+    int fieldCount(size_t dataSourceLayerNum = 0) const;
     /**
      * Select features based on a bounding rectangle. Features can be retrieved 
      * with calls to getFirstFeature and getNextFeature.
      * @param mbr QgsRect containing the extent to use in selecting features
      * @param useIntersect Use geos functions to determine the selected set
      */
-    void select(QgsRect * mbr, bool useIntersect = false, int dataSourceLayerNum = 0);
+    void select(QgsRect * mbr, bool useIntersect = false, size_t dataSourceLayerNum = 0);
 
     /**
      * Identify features within the search radius specified by rect
@@ -124,49 +124,49 @@ class QgsOgrProvider:public QgsVectorDataProvider
      * @param rect Bounding rectangle of search radius
      * @return std::vector containing QgsFeature objects that intersect rect
      */
-    //virtual std::vector < QgsFeature > &identify(QgsRect * rect, int dataSourceLayerNum = 0);
-    void identify(QgsRect * rect, int dataSourceLayerNum = 0);
+    //virtual std::vector < QgsFeature > &identify(QgsRect * rect, size_t dataSourceLayerNum = 0);
+    void identify(QgsRect * rect, size_t dataSourceLayerNum = 0);
 
     /** Return the extent for this data layer
     */
-    virtual QgsRect *extent( int dataSourceLayerNum = 0 );
+    virtual QgsRect *extent( size_t dataSourceLayerNum = 0 );
 
     /**Get an attribute associated with a feature*/
-    void getFeatureAttribute(OGRFeature * ogrFet, QgsFeature * f, int attindex, int dataSourceLayerNum = 0);
+    void getFeatureAttribute(OGRFeature * ogrFet, QgsFeature * f, int attindex, size_t dataSourceLayerNum = 0);
     /**
      * Get the attributes associated with a feature
      */
-    void getFeatureAttributes(OGRFeature * ogrFet, QgsFeature * f, int dataSourceLayerNum = 0);
+    void getFeatureAttributes(OGRFeature * ogrFet, QgsFeature * f, size_t dataSourceLayerNum = 0);
     /**
      * Get the field information for the layer
      */
-    std::vector < QgsField > const & fields(int dataSourceLayerNum = 0) const;
+    std::vector < QgsField > const & fields(size_t dataSourceLayerNum = 0) const;
 
     /* Reset the layer - for an OGRLayer, this means clearing the
      * spatial filter and calling ResetReading
      */
-    /* virtual */ void reset(int dataSourceLayerNum = 0);
+    /* virtual */ void reset(size_t dataSourceLayerNum = 0);
 
     /**Returns the minimum value of an attribut
       @param position the number of the attribute*/
-    QString minValue(int position, int dataSourceLayerNum = 0);
+    QString minValue(int position, size_t dataSourceLayerNum = 0);
 
     /**Returns the maximum value of an attribut
       @param position the number of the attribute*/
-    QString maxValue(int position, int dataSourceLayerNum = 0);
+    QString maxValue(int position, size_t dataSourceLayerNum = 0);
 
     /**Returns true if this is a valid shapefile
     */
     bool isValid() const;
 
     /**Writes a list of features to the file*/
-    bool addFeatures(std::list<QgsFeature*> const flist, int dataSourceLayerNum = 0);
+    bool addFeatures(std::list<QgsFeature*> const flist, size_t dataSourceLayerNum = 0);
 
     /**Adds new attributess. Unfortunately not supported for layers with features in it*/
-    bool addAttributes(std::map<QString,QString> const & name, int dataSourceLayerNum = 0);
+    bool addAttributes(std::map<QString,QString> const & name, size_t dataSourceLayerNum = 0);
 
     /**Changes attribute values of existing features */
-    bool changeAttributeValues(std::map<int,std::map<QString,QString> > const & attr_map, int dataSourceLayerNum = 0);
+    bool changeAttributeValues(std::map<int,std::map<QString,QString> > const & attr_map, size_t dataSourceLayerNum = 0);
 
     QgsDataSourceURI * getURI()
     { 
@@ -175,7 +175,7 @@ class QgsOgrProvider:public QgsVectorDataProvider
 
     /**Tries to create a .qix index file for faster access if only a subset of the features is required
      @return true in case of success*/
-    bool createSpatialIndex(int dataSourceLayerNum = 0);
+    bool createSpatialIndex(size_t dataSourceLayerNum = 0);
 
     /** Returns a bitmask containing the supported capabilities
         Note, some capabilities may change depending on whether
@@ -253,10 +253,10 @@ class QgsOgrProvider:public QgsVectorDataProvider
   private:
 
     /**Adds one feature*/
-    bool addFeature(QgsFeature* f, int dataSourceLayerNum = 0);
+    bool addFeature(QgsFeature* f, size_t dataSourceLayerNum = 0);
 
     /**Fills the cash and sets minmaxcachedirty to false*/
-    void fillMinMaxCache(int dataSourceLayerNum = 0);
+    void fillMinMaxCache(size_t dataSourceLayerNum = 0);
 
     // XXX isn't this already ultimately defined in QgsDataProvider?
     // XXX (commented out to see if it is so)
